# perf: Optimize discovery.relabel by using label builder